chore: desacoplar dependencias y renderizado de dev-tools
